+++ src/capsula-container.bash
@@ -99,7 +99,7 @@ fi
 if ! getent passwd $usr >/dev/null 2>&1; then
     if [[ $platform == "alpine" ]]; then
         useradd -M -d $homedir -s $SHELL -u $uid -g $grp -G wheel $usr >/dev/null 2>&1
-    elif [[ $platform == "debian" ]]; then
+    elif [[ $platform == "debian" || $platform == "ubuntu" ]]; then
         useradd -M -d $homedir -s $SHELL -u $uid -g $grp -G sudo $usr
     fi
 fi